+++ jeesns-group/jeesns-group-web/src/main/java/com/lxinet/jeesns/group/web/manage/GroupController.java
@@ -1,6 +1,7 @@
 package com.lxinet.jeesns.group.web.manage;
 
 import com.lxinet.jeesns.commons.utils.MemberUtil;
+import com.lxinet.jeesns.core.model.MemberToken;
 import com.lxinet.jeesns.member.interceptor.AdminLoginInterceptor;
 import com.lxinet.jeesns.core.annotation.Before;
 import com.lxinet.jeesns.core.dto.ResponseModel;
@@ -8,6 +9,7 @@ import com.lxinet.jeesns.core.model.Page;
 import com.lxinet.jeesns.group.service.IGroupService;
 import com.lxinet.jeesns.core.web.BaseController;
 import com.lxinet.jeesns.member.model.Member;
+import com.lxinet.jeesns.member.service.IMemberService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -18,34 +20,43 @@ import javax.annotation.Resource;
  */
 @Controller("manageGroupController")
 @RequestMapping("/")
-@Before(AdminLoginInterceptor.class)
 public class GroupController extends BaseController {
-    private static final String MANAGE_FTL_PATH = "/manage/group/";
     @Resource
     private IGroupService groupService;
+    @Resource
+    private IMemberService memberService;
 
-    @RequestMapping(value = "${managePath}/group/index")
-    public String index(@RequestParam(value = "status",required = false,defaultValue = "-1") Integer status,
-                        String key,
-                        Model model) {
+    @RequestMapping(value = "${managePath}/group/list")
+    @ResponseBody
+    public Object list(@RequestParam(value = "status",required = false,defaultValue = "-1") Integer status, String key) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         Page page = new Page(request);
         ResponseModel responseModel = groupService.listByPage(status,page,key);
-        model.addAttribute("model",responseModel);
-        model.addAttribute("key",key);
-        return MANAGE_FTL_PATH + "index";
+        return responseModel;
     }
 
-    @RequestMapping(value = "${managePath}/group/delete/{id}",method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/group/delete",method = RequestMethod.GET)
     @ResponseBody
-    public Object delete(@PathVariable("id") int id){
-        Member loginMember = MemberUtil.getLoginMember(request);
+    public Object delete(@RequestParam("id") int id){
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         ResponseModel response = groupService.delete(loginMember,id);
         return response;
     }
 
-    @RequestMapping(value = "${managePath}/group/changeStatus/{id}",method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/group/changeStatus",method = RequestMethod.GET)
     @ResponseBody
-    public Object changeStatus(@PathVariable("id") int id){
+    public Object changeStatus(@RequestParam("id") int id){
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         ResponseModel response = groupService.changeStatus(id);
         return response;
     }
