+++ jeesns-weibo/jeesns-weibo-web/src/main/java/com/lxinet/jeesns/weibo/web/manage/WeiboController.java
@@ -1,17 +1,15 @@
 package com.lxinet.jeesns.weibo.web.manage;
 
-import com.lxinet.jeesns.commons.utils.MemberUtil;
-import com.lxinet.jeesns.member.interceptor.AdminLoginInterceptor;
-import com.lxinet.jeesns.core.annotation.Before;
+import com.lxinet.jeesns.core.model.MemberToken;
 import com.lxinet.jeesns.core.dto.ResponseModel;
 import com.lxinet.jeesns.core.model.Page;
 import com.lxinet.jeesns.core.web.BaseController;
 import com.lxinet.jeesns.member.model.Member;
+import com.lxinet.jeesns.member.service.IMemberService;
 import com.lxinet.jeesns.weibo.service.IWeiboService;
+import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
 import javax.annotation.Resource;
 
 /**
@@ -19,24 +17,32 @@ import javax.annotation.Resource;
  */
 @Controller("mamageWeiboController")
 @RequestMapping("/")
-@Before(AdminLoginInterceptor.class)
 public class WeiboController extends BaseController {
-    private static final String MANAGE_FTL_PATH = "/manage/weibo/";
     @Resource
     private IWeiboService weiboService;
+    @Resource
+    private IMemberService memberService;
 
-    @RequestMapping("${managePath}/weibo/index")
-    public String index(@RequestParam(value = "key",required = false,defaultValue = "") String key, Model model) {
+    @RequestMapping("${managePath}/weibo/list")
+    @ResponseBody
+    public Object list(@RequestParam(value = "key",required = false,defaultValue = "") String key) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
         Page page = new Page(request);
         ResponseModel responseModel = weiboService.listByPage(page,0,0,key);
-        model.addAttribute("model",responseModel);
-        return MANAGE_FTL_PATH + "index";
+        return responseModel;
     }
 
-    @RequestMapping(value = "${managePath}/weibo/delete/{id}", method = RequestMethod.GET)
+    @RequestMapping(value = "${managePath}/weibo/delete", method = RequestMethod.GET)
     @ResponseBody
-    public Object delete(@PathVariable("id") int id) {
-        Member loginMember = MemberUtil.getLoginMember(request);
+    public Object delete(@Param("id") int id) {
+        ResponseModel<MemberToken> validMemberTokenModel = validMemberToken();
+        if (validMemberTokenModel.getCode() == -1){
+            return validMemberTokenModel;
+        }
+        Member loginMember = memberService.findById(validMemberTokenModel.getData().getMemberId());
         return weiboService.delete(request, loginMember,id);
     }
 }
